+++ client/src/pages/index.tsx
@@ -6,10 +6,17 @@ const Home: React.FC = () => (
     <Head>
       <title>Welcome to Landgriffon</title>
     </Head>
-    <ApplicationLayout>
-      <p>Welcome</p>
-    </ApplicationLayout>
+    <ApplicationLayout>{/* Welcome */}</ApplicationLayout>
   </>
 );
+
+export async function getServerSideProps() {
+  return {
+    redirect: {
+      destination: '/analysis',
+      permanent: false,
+    },
+  };
+}
 
 export default Home;